+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Technically, it's still not required, but there will no longer be any warnings if it's not available.
 - `DieFoundry` no longer uses Symbol for string sides, they are now Strings.
   - This coincides with changes in **vector_number** 0.7.0 which removed special treatment for String units and unified behavior.
+- "-v" option is now used for "--verbose", enabling extra output.
 
 [Compare v0.17.1...main](https://github.com/trinistr/dicey/compare/v0.17.1...main)